requirements.txt
opencv-python>=4.5
numpy
pyyaml
face_recognition
imutils
pickle5
flask
flask-cors
pillow
requests
